Move QuantumStateRepresentation to its own file
Fixes: #5475
