+++ cirq-core/cirq/qis/__init__.py
@@ -25,9 +25,11 @@ from cirq.qis.channels import (
     superoperator_to_kraus,
 )
 
-from cirq.qis.clifford_tableau import CliffordTableau, QuantumStateRepresentation, StabilizerState
+from cirq.qis.clifford_tableau import CliffordTableau, StabilizerState
 
 from cirq.qis.measures import entanglement_fidelity, fidelity, von_neumann_entropy
+
+from cirq.qis.quantum_state_representation import QuantumStateRepresentation
 
 from cirq.qis.states import (
     bloch_vector_from_state_vector,
